完成了WorkerService，同时根据实现调整了几处mapper，Dao，DTO

+++ src/main/java/team/combinatorics/midtermproject/model/dto/AllWorkerDTO.java
@@ -14,5 +14,5 @@ import java.util.List;
 public class AllWorkerDTO {
     private List<WorkerDTO> allWorker;
     private Integer allWorkerNum;
-    private Float allWorkerAvgSalary;
+    private Double allWorkerAvgSalary;
 }
